Removed Example, added new script for use with any number of objects
